Refactor CLI command names PR comment fixes
