+++ src/SprykerSdk/Zed/AppSdk/Communication/Console/AppTranslationValidateConsole.php
@@ -15,26 +15,28 @@ use Symfony\Component\Console\Output\OutputInterface;
 /**
  * @method \SprykerSdk\Zed\AppSdk\Business\AppSdkFacadeInterface getFacade()
  */
-class ConfigurationValidateConsole extends AbstractConsole
+class AppTranslationValidateConsole extends AbstractConsole
 {
     /**
      * @var string
      */
-    public const CONFIGURATION_FILE = 'configuration-file';
+    public const TRANSLATION_FILE = 'translation-file';
 
     /**
      * @var string
      */
-    public const CONFIGURATION_FILE_SHORT = 'c';
+    public const TRANSLATION_FILE_SHORT = 't';
 
     /**
      * @return void
      */
     protected function configure(): void
     {
-        $this->setName('app:configuration:validate')
-            ->setDescription('Validates the configuration file.')
-            ->addOption(static::CONFIGURATION_FILE, static::CONFIGURATION_FILE_SHORT, InputOption::VALUE_REQUIRED, '', $this->getConfig()->getDefaultConfigurationFile());
+        $this->setName('app:translation:validate')
+            ->setDescription('Validates the translation file.')
+            ->addOption(static::TRANSLATION_FILE, static::TRANSLATION_FILE_SHORT, InputOption::VALUE_REQUIRED, '', $this->getConfig()->getDefaultTranslationFile())
+            ->addOption(AppManifestValidateConsole::MANIFEST_PATH, AppManifestValidateConsole::MANIFEST_PATH_SHORT, InputOption::VALUE_REQUIRED, '', $this->getConfig()->getDefaultManifestPath())
+            ->addOption(AppConfigurationValidateConsole::CONFIGURATION_FILE, AppConfigurationValidateConsole::CONFIGURATION_FILE_SHORT, InputOption::VALUE_REQUIRED, '', $this->getConfig()->getDefaultConfigurationFile());
     }
 
     /**
@@ -46,9 +48,11 @@ class ConfigurationValidateConsole extends AbstractConsole
     public function execute(InputInterface $input, OutputInterface $output): int
     {
         $validateRequestTransfer = new ValidateRequestTransfer();
-        $validateRequestTransfer->setConfigurationFile($input->getOption(static::CONFIGURATION_FILE));
+        $validateRequestTransfer->setTranslationFile($input->getOption(static::TRANSLATION_FILE));
+        $validateRequestTransfer->setManifestPath($input->getOption(AppManifestValidateConsole::MANIFEST_PATH));
+        $validateRequestTransfer->setConfigurationFile($input->getOption(AppConfigurationValidateConsole::CONFIGURATION_FILE));
 
-        $validateResponseTransfer = $this->getFacade()->validateConfiguration($validateRequestTransfer);
+        $validateResponseTransfer = $this->getFacade()->validateTranslation($validateRequestTransfer);
 
         if ($validateResponseTransfer->getErrors()->count() === 0) {
             return static::CODE_SUCCESS;
