+++ tests/SprykerSdkTest/Zed/AppSdk/Communication/Console/AppManifestValidateConsoleTest.php
@@ -9,7 +9,7 @@ namespace SprykerSdkTest\Zed\AppSdk\Communication\Console;
 
 use Codeception\Test\Unit;
 use SprykerSdk\Zed\AppSdk\Communication\Console\AbstractConsole;
-use SprykerSdk\Zed\AppSdk\Communication\Console\AsyncApiValidateConsole;
+use SprykerSdk\Zed\AppSdk\Communication\Console\AppManifestValidateConsole;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
@@ -18,9 +18,9 @@ use Symfony\Component\Console\Output\OutputInterface;
  * @group AppSdk
  * @group Communication
  * @group Console
- * @group ValidateAsyncApiConsoleTest
+ * @group AppManifestValidateConsoleTest
  */
-class ValidateAsyncApiConsoleTest extends Unit
+class AppManifestValidateConsoleTest extends Unit
 {
     /**
      * @var \SprykerSdkTest\Zed\AppSdk\CommunicationTester
@@ -30,12 +30,12 @@ class ValidateAsyncApiConsoleTest extends Unit
     /**
      * @return void
      */
-    public function testValidateAsyncApiReturnsSuccessCodeWhenValidationIsSuccessful(): void
+    public function testValidateManifestReturnsSuccessCodeWhenValidationIsSuccessful(): void
     {
         // Arrange
-        $this->tester->haveValidAsyncApiFile();
+        $this->tester->haveValidManifestFile();
 
-        $commandTester = $this->tester->getConsoleTester(AsyncApiValidateConsole::class);
+        $commandTester = $this->tester->getConsoleTester(AppManifestValidateConsole::class);
 
         // Act
         $commandTester->execute([]);
@@ -47,10 +47,10 @@ class ValidateAsyncApiConsoleTest extends Unit
     /**
      * @return void
      */
-    public function testValidateAsyncApiReturnsErrorCodeAndPrintsErrorMessagesWhenValidationFailed(): void
+    public function testValidateManifestReturnsErrorCodeAndPrintsErrorMessagesWhenValidationFailed(): void
     {
         // Arrange
-        $commandTester = $this->tester->getConsoleTester(new AsyncApiValidateConsole());
+        $commandTester = $this->tester->getConsoleTester(new AppManifestValidateConsole());
 
         // Act
         $commandTester->execute([], ['verbosity' => OutputInterface::VERBOSITY_VERBOSE]);
